+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.3
 - Add option to disable safety settings
+- Add option to exclude locales
 
 ## 0.1.2
 - Fix setting model provider in `l10n.yaml`